Reddit_Data_Collection/requirements.txt
asyncpraw==7.7.1
pandas==2.2.3
aiohttp==3.9.3
python-dotenv==0.21.0
tqdm==4.66.2